chore: TokenCookieProvider 네이밍 변경
